inpaint/requirements.txt
torch>=1.12.0
torchvision>=0.13.0
diffusers>=0.21.0
transformers>=4.21.0
accelerate>=0.20.0
gradio>=4.0.0
Pillow>=9.0.0
numpy>=1.21.0
safetensors>=0.3.0
xformers>=0.0.20